Feat: #TW-157 - uWSGI 설정하기
uWSGI 설정파일 생성
Flask에서 uWSGI로 서버가 넘어감에 따라 실행코드 수정
