+++ ai/KcELECTRAchatbot/flask/main.py
@@ -323,4 +323,4 @@ def action2():
 
 # Run the app
 if __name__ == '__main__':
-    app.run(host='0.0.0.0', debug=True, port=8000)
+    app.run(host='0.0.0.0', port=8000)
