+++ ai/KcELECTRAchatbot/flask/uwsgi.ini
@@ -1,0 +1,9 @@
+[uwsgi]
+http = :8000
+
+master = true
+wsgi-file = main.py
+callable = app
+
+vacuum = true
+die-on-term = true
