+++ ai/KoSentenceBERTchatbot/uwsgi.ini
@@ -1,0 +1,11 @@
+[uwsgi]
+http = :8080
+
+master = true
+wsgi-file = main.py
+callable = app
+
+vacuum = true
+die-on-term = true
+
+venv = '/opt/conda'
